Move PyInterpreterState to the internal header file.
