+++ Modules/_testcapimodule.c
@@ -4685,7 +4685,7 @@ static PyObject *
 get_core_config(PyObject *self, PyObject *Py_UNUSED(args))
 {
     PyInterpreterState *interp = _PyInterpreterState_Get();
-    const _PyCoreConfig *config = &interp->core_config;
+    const _PyCoreConfig *config = _PyInterpreterState_GetCoreConfig(interp);
     return _PyCoreConfig_AsDict(config);
 }
 
@@ -4694,7 +4694,7 @@ static PyObject *
 get_main_config(PyObject *self, PyObject *Py_UNUSED(args))
 {
     PyInterpreterState *interp = _PyInterpreterState_Get();
-    const _PyMainInterpreterConfig *config = &interp->config;
+    const _PyMainInterpreterConfig *config = _PyInterpreterState_GetMainConfig(interp);
     return _PyMainInterpreterConfig_AsDict(config);
 }
